requirements.txt
certifi==2024.8.30
psycopg2-binary
selenium==4.27.1
trio-websocket==0.11
trio==0.27.0
typing-extensions==4.12
urllib3[socks]==2.2
websocket-client==1.8
requests==2.32.3
python-telegram-bot
python-dotenv